requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic>=2.12,<3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
alembic==1.13.2
psycopg[binary]==3.2.1

geoalchemy2>=0.14.6

email-validator==2.2.0
psycopg2-binary==2.9.9
requests

bcrypt==4.1.2
